refactor(species-id): move IdentifyRequest into shared protocol crate
The /identify request body was defined twice — IdentifyRequest in the
service's types module and IdentifyRequestBody in the appview client —
describing the same wire shape. Move it into the protocol crate
alongside IdentifyResponse and SpeciesSuggestion (where the wire
contract belongs) and re-export through the existing locations so
consumers don't break.
